requirements.txt
asgiref==3.5.0
attrs==21.4.0
cached-property==1.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0.4
django-admin-interface==0.19.1
django-admin-volt==1.0.8
django-colorfield==0.7.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-jalali==6.0.1
django-jalali-date==1.0.1
django-modeltranslation==0.18.10
django-tinymce==3.5.0
django-user-agents==0.4.0
idna==3.4
isodate==0.6.1
jdatetime==4.1.0
lxml==4.9.1
mysqlclient==2.1.0
Pillow==9.1.1
platformdirs==2.5.2
python-decouple==3.1
pytz==2022.1
requests==2.28.1
requests-file==1.5.1
requests-toolbelt==0.9.1
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.6.3
tzdata==2022.1
ua-parser==0.16.1
Unidecode==1.3.6
urllib3==1.26.10
user-agents==2.2.0
zeep==4.1.0
